backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0
numpy==1.26.3

# Report generation
reportlab==4.0.9
Jinja2==3.1.3

# Git operations
GitPython==3.1.41

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2

# Optional: For model loading if needed
joblib==1.3.2
scikit-learn==1.4.0

# Groq API for LLM integration
groq==0.4.2